Pr feedback move DTOs to model package same as level as api
