+++ extensions/dataspace-authority-spi/src/testFixtures/java/org/eclipse/dataspaceconnector/registration/TestUtils.java
@@ -16,12 +16,9 @@ package org.eclipse.dataspaceconnector.registration;
 
 import com.github.javafaker.Faker;
 import org.eclipse.dataspaceconnector.registration.authority.model.Participant;
-import org.eclipse.dataspaceconnector.registration.authority.model.ParticipantDto;
 import org.eclipse.dataspaceconnector.registration.authority.model.ParticipantStatus;
-import org.eclipse.dataspaceconnector.registration.authority.model.ParticipantStatusDto;
 
 import java.util.List;
-import java.util.Map;
 
 import static java.lang.String.format;
 
@@ -39,41 +36,5 @@ public class TestUtils {
                 .url(FAKER.internet().url())
                 .supportedProtocols(List.of(FAKER.lorem().word(), FAKER.lorem().word()))
                 .supportedProtocol(FAKER.lorem().word());
-    }
-
-    public static ParticipantDto.Builder createParticipantDto() {
-        return ParticipantDto.Builder.newInstance()
-                .did(format("did:web:%s", FAKER.internet().domainName()))
-                .status(FAKER.options().option(ParticipantStatusDto.class))
-                .name(FAKER.lorem().characters())
-                .url(FAKER.internet().url())
-                .supportedProtocols(List.of(FAKER.lorem().word(), FAKER.lorem().word()))
-                .supportedProtocol(FAKER.lorem().word());
-
-    }
-
-    public static ParticipantDto getParticipantDtoFromParticipant(Participant participant) {
-        return ParticipantDto.Builder.newInstance()
-                .did(participant.getDid())
-                .status(modelToDtoStatusMap().get(participant.getStatus()))
-                .name(participant.getName())
-                .url(participant.getUrl())
-                .supportedProtocols(participant.getSupportedProtocols())
-                .build();
-    }
-
-    /**
-     * Map of ParticipantStatus & ParticipantStatusDto.
-     * It describes what should be DTO status in respect of domain model status.
-     */
-    private static Map<ParticipantStatus, ParticipantStatusDto> modelToDtoStatusMap() {
-        return Map.of(
-                ParticipantStatus.ONBOARDING_INITIATED, ParticipantStatusDto.ONBOARDING_IN_PROGRESS,
-                ParticipantStatus.AUTHORIZING, ParticipantStatusDto.ONBOARDING_IN_PROGRESS,
-                ParticipantStatus.AUTHORIZED, ParticipantStatusDto.ONBOARDING_IN_PROGRESS,
-                ParticipantStatus.ONBOARDED, ParticipantStatusDto.ONBOARDED,
-                ParticipantStatus.FAILED, ParticipantStatusDto.DENIED,
-                ParticipantStatus.DENIED, ParticipantStatusDto.DENIED
-        );
     }
 }
