+++ extensions/registration-service/src/main/java/org/eclipse/dataspaceconnector/registration/model/ParticipantDto.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.registration.authority.model;
+package org.eclipse.dataspaceconnector.registration.model;
 
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import io.swagger.v3.oas.annotations.media.Schema;
